Try and buffer the remote signal before kicking off decoding. This fixes timing issues with signals overlapping 2 audio buffers.
git-svn-id: svn+ssh://svn.code.sf.net/p/winlirc/code/trunk@223 92c2518b-485e-4ca2-9c37-138eef3f838b

+++ DLL/AudioCapture/StringFunctions.cpp
@@ -2,9 +2,9 @@
 
 void removeTrailingWhiteSpace(TCHAR *string) {
 
-	//===============
-	int stringLength;
-	//===============
+	//==================
+	size_t stringLength;
+	//==================
 
 	//
 	// get string length
